cast bool mask to float32 for op

+++ spektral/layers/convolutional/ecc_conv.py
@@ -182,7 +182,7 @@ class ECCConv(Conv):
 
         # this doesn't work as expected (passing `None` to `mask` doesn't make it None. It makes it [None]*3, which makes it think that there's a mask.
         if mask is not None or len(tf.where(mask != None)) > 0:
-            output *= mask[0]
+            output *= tf.cast(mask[0], tf.float32)
         output = self.activation(output)
 
         return output
